Heartbeat impl fixed. Independant Mains which launch RM and GS.
The problem with the HB was the synchronization and to that I got wrong
the idea. Now it's working properly. The key is to implement the
HeartbeatReceiver interface (before called Sender) by RM and GS.

Created separated mains for launching RM and GS. CHANGE THE FILEPATH on
the SimulationLauncherFile before launching it!!! I was getting
exceptios because it didn't find the rm/rms file so I hardcoded it.
Eventually it will be a path to a S3 bucket in AWS.

There is another simulation runner which shows how the HB systems works
when you shutdown a RM and restart it.